Added - images for rural and city flights, routes_forms.py, getting_started.html
Updated -
_init_.py
routes.py,
routes_wizard.py,
viabilityStudy.html
