+++ app/routes.py
@@ -1,7 +1,6 @@
 from app import app
-from flask import render_template, request
+from flask import render_template
 
-from app.forms.viabilityStudyForm import ViabilityStudyForm
 from app.models import User, Role
 
 
@@ -23,19 +22,5 @@ def terms_of_service():
 @app.route("/contact-us")
 def contact_us():
     return render_template('/info/contact_us.html', title='Contact')
-@app.route('/viability-study', methods=['GET', 'POST'])
-def viability_study():
-    form = ViabilityStudyForm()
-
-    if request.method == "POST":
-        # saves changes only for finalising later
-        if form.saveChanges.data:
-            print("Form Saved")
-
-        # submits completed form
-        if form.validate_on_submit() and form.submit.data:
-            print("Form Submitted")
-
-    return render_template("forms/viabilityStudy.html", form=form)
