+++ app/routes_forms.py
@@ -1,0 +1,18 @@
+from app import app
+from flask import render_template, request
+from app.forms.viabilityStudyForm import ViabilityStudyForm
+
+@app.route('/forms/viability-study', methods=['GET', 'POST'])
+def viability_study():
+    form = ViabilityStudyForm()
+
+    if request.method == "POST":
+        # saves changes only for finalising later
+        if form.saveChanges.data:
+            print("Form Saved")
+
+        # submits completed form
+        if form.submit.data and form.validate_on_submit():
+            print("Form Submitted")
+
+    return render_template("/forms/viabilityStudy.html", form=form)
